Сплошной код поделен на модули по задачам
